refactory model package
