+++ models/user.go
@@ -1,4 +1,4 @@
-package model
+package models
 
 import "go.mongodb.org/mongo-driver/bson/primitive"
